server/requirements.txt
fastapi
uvicorn
python-socketio
sqlalchemy
pydantic
python-multipart
pytest
httpx
